pyproject.toml
[project]
name = "hustle4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "crewai>=0.159.0",
    "crewai-tools>=0.62.0",
    "cryptography>=45.0.6",
    "firebase-admin>=7.1.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "google-genai>=1.30.0",
    "gunicorn>=23.0.0",
    "openai>=1.99.9",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.4",
    "svix>=1.73.0",
]
